py/requirements.txt
grpcio==1.42.0
lightgbm==3.3.1
numpy==1.21.4
pandas==1.3.4
protobuf==3.19.1
tensorflow==2.8.0rc0
